requirements.txt
Flask==1.1.2
gunicorn==20.0.4
h5py==2.10.0
Jinja2==2.11.2
numpy==1.18.3
Pillow==7.1.2
tensorflow==2.0.0
Werkzeug==1.0.1

